[kernel] uprobes: Cleanup the usage of arch_uprobe->def.fixups, make it u8
Message-id: <20140616160211.GA15507@redhat.com>
Patchwork-id: 84091
O-Subject: [PATCH RHEL7 BZ 1073627 27/36] uprobes/x86: Cleanup the usage of arch_uprobe->def.fixups, make it u8
Bugzilla: 1073627
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

RHBZ: 1073627

handle_riprel_insn() assumes that nobody else could modify ->fixups
before. This is correct but fragile, change it to use "|=".

Also make ->fixups u8, we are going to add the new members into the
union. It is not clear why UPROBE_FIX_RIP_.X lived in the upper byte,
redefine them so that they can fit into u8.

(cherry picked from commit 78d9af4cd375880a574327210eb9dab572618364)

+++ arch/x86/include/asm/uprobes.h
@@ -50,7 +50,7 @@ struct arch_uprobe {
 #ifdef CONFIG_X86_64
 			long	riprel_target;
 #endif
-			u16	fixups;
+			u8	fixups;
 		} 			def;
 	};
 };
